refactor: better class and function names
